Create form with ajax, I've convert the primary key of Factura on a bigIncrements

+++ database/migrations/2022_02_28_085032_create_lineas_table.php
@@ -19,7 +19,7 @@ class CreateLineasTable extends Migration
             $table->string('producto');
             $table->float('cantidad')->default(1);
             $table->float('precio')->default(0);
-            $table->integer('factura_numero');
+            $table->unsignedBigInteger('factura_numero');
             $table->foreign('factura_numero')
                 ->references('numero')
                 ->on('facturas')
